refactor: :recycle: split in Clean DDD
