+++ src/App.tsx
@@ -1,14 +1,15 @@
 import CssBaseline from '@mui/material/CssBaseline';
-import Header from './components/header';
-import Home from './pages/home';
-import Books from './pages/books';
-import TabletopGames from './pages/tabletop-game';
-import AboutMe from './pages/about-me';
-import NotFound from './pages/not-found';
-import React from 'react';
+import Header from './core/components/header';
+import Home from './core/views/home';
+import React, { lazy, Suspense } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { createTheme, ThemeProvider } from '@mui/material';
 import { useTranslation } from 'react-i18next';
+
+const Books = lazy(() => import('./books/view/books'));
+const TabletopGames = lazy(() => import('./tabletopGame/view/tabletop-game'));
+const AboutMe = lazy(() => import('./aboutMe/view/about-me'));
+const NotFound = lazy(() => import('./core/views/not-found'));
 
 export const ColorModeContext = React.createContext({
   // eslint-disable-next-line @typescript-eslint/no-empty-function
@@ -50,13 +51,15 @@ function App(): JSX.Element {
         <CssBaseline />
         <Header title={t<string>('title')} />
         <BrowserRouter>
-          <Routes>
-            <Route path={routes.home} element={<Home />} />
-            <Route path={routes.book} element={<Books />} />
-            <Route path={routes.tabletopGame} element={<TabletopGames />} />
-            <Route path={routes.aboutMe} element={<AboutMe />} />
-            <Route path="/*" element={<NotFound />} />
-          </Routes>
+          <Suspense fallback={<div>Loading...</div>}>
+            <Routes>
+              <Route path={routes.home} element={<Home />} />
+              <Route path={routes.book} element={<Books />} />
+              <Route path={routes.tabletopGame} element={<TabletopGames />} />
+              <Route path={routes.aboutMe} element={<AboutMe />} />
+              <Route path="/*" element={<NotFound />} />
+            </Routes>
+          </Suspense>
         </BrowserRouter>
       </ThemeProvider>
     </ColorModeContext.Provider>
