+++ src/aboutMe/components/skill-container.tsx
@@ -5,7 +5,7 @@ import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
 import { FC, useState } from 'react';
-import { SkillContainr } from '../types';
+import { SkillContainr } from '../../types';
 import SkillItem from './skill-item';
 
 const SkillContainer: FC<SkillContainr> = (props: SkillContainr) => {
