+++ src/aboutMe/entities/certification.tsx
@@ -1,0 +1,6 @@
+export type Certification = {
+    certificationCompany: string;
+    logo: string;
+    name: string;
+    obtentionDate: string;
+};
