+++ src/aboutMe/entities/experience.tsx
@@ -1,0 +1,11 @@
+type Experience = {
+    company: string;
+    companyLogo: string;
+    context: string;
+    position: string;
+    tasks: string[];
+    tags: string[];
+    timeRanges: string;
+};
+
+export default Experience;
