+++ src/core/views/home.tsx
