+++ src/types.tsx
@@ -1,20 +1,3 @@
-export type Certification = {
-  certificationCompany: string;
-  logo: string;
-  name: string;
-  obtentionDate: string;
-};
-
-export type Experience = {
-  company: string;
-  companyLogo: string;
-  context: string;
-  position: string;
-  tasks: string[];
-  tags: string[];
-  timeRanges: string;
-};
-
 export type Hobby = {
   abbreviation: string;
   federationLogo: string;
